Move OptionExerciseModule to models folder
